Cover Metrics

This section provides metrics related to coral cover. These metrics calculate coral cover at different spatial scales and for different taxonomic groups.

ADRIAIndicators.ltmp_cover! — Method

ltmp_cover!(relative_cover::AbstractArray{T,4}, habitable_area::AbstractVector{T}, reef_area::AbstractVector{T}, out_ltmp_cover::AbstractArray{T,2})::Nothing where {T<:Real}

Calculate the LTMP cover for each location. LTMP cover is the proportion of the reef area occupied by coral relative to the area of the reef, which may include non-habitable area. The relative_cover input is relative to habitable area.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

habitable_area : The habitable area for each location in m².

reef_area : The total area for each location in m².

out_ltmp_cover : Output array buffer for LTMP cover with dimensions [timesteps ⋅ locations].

ADRIAIndicators.ltmp_cover — Method

ltmp_cover(relative_cover::AbstractArray{T,4}, habitable_area::AbstractVector{T}, reef_area::AbstractVector{T})::AbstractArray{T,2} where {T<:Real}

Calculate the LTMP cover for each location. LTMP cover is the proportion of the reef area occupied by coral relative to the area of the reef, which may include non-habitable area. The relative_cover input is relative to habitable area.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

habitable_area : The habitable area for each location in m².

reef_area : The total area for each location in m².

Returns

A 2D array of LTMP cover with dimensions [timesteps ⋅ locations].

ADRIAIndicators.ltmp_loc_taxa_cover! — Method

ltmp_loc_taxa_cover!(relative_cover::AbstractArray{T,4}, habitable_area::AbstractVector{T}, reef_area::AbstractVector{T}, out_ltmp_loc_taxa_cover::AbstractArray{T,3})::Nothing where {T<:Real}

Calculate the LTMP taxa cover for each location.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

habitable_area : The habitable area for each location in m².

reef_area : The total area for each location in m².

out_ltmp_loc_taxa_cover : Output array buffer with dimensions [timesteps ⋅ groups ⋅ locations].

ADRIAIndicators.ltmp_loc_taxa_cover — Method

ltmp_loc_taxa_cover(relative_cover::AbstractArray{T,4}, habitable_area::AbstractVector{T}, reef_area::AbstractVector{T})::AbstractArray{T,3} where {T<:Real}

Calculate the LTMP taxa cover for each location.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

habitable_area : The habitable area for each location in m².

reef_area : The total area for each location in m².

Returns

A 3D array of LTMP taxa cover with dimensions [timesteps ⋅ groups ⋅ locations].

ADRIAIndicators.ltmp_taxa_cover! — Method

ltmp_taxa_cover!(relative_cover::AbstractArray{T,4}, habitable_area::AbstractVector{T}, reef_area::AbstractVector{T}, out_ltmp_taxa_cover::AbstractArray{T,2})::Nothing where {T<:Real}

Calculate the LTMP taxa cover, summed up across all locations.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

habitable_area : The habitable area for each location in m².

reef_area : The total area for each location in m².

out_ltmp_taxa_cover : Output array buffer with dimensions [timesteps ⋅ groups].

ADRIAIndicators.ltmp_taxa_cover — Method

ltmp_taxa_cover(relative_cover::AbstractArray{T,4}, habitable_area::AbstractVector{T}, reef_area::AbstractVector{T})::AbstractArray{T,2} where {T<:Real}

Calculate the LTMP taxa cover, summed up across all locations.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

habitable_area : The habitable area for each location in m².

reef_area : The total area for each location in m².

Returns

A 2D array of LTMP taxa cover with dimensions [timesteps ⋅ groups].

ADRIAIndicators.relative_cover! — Method

relative_cover!(relative_cover::AbstractArray{T,4}, out_relative_cover::AbstractArray{T,2})::Nothing where {T<:Real}

Calculate the relative cover per location by summing over groups and size classes.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

out_relative_cover : Output array buffer for relative cover with dimensions [timesteps ⋅ locations].

ADRIAIndicators.relative_cover — Method

relative_cover(relative_cover::AbstractArray{T,4})::AbstractArray{T,2} where {T<:Real}

Calculate the relative cover per location by summing over groups and size classes.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

Returns

A 2D array of relative cover with dimensions [timesteps ⋅ locations].

ADRIAIndicators.relative_loc_taxa_cover! — Method

relative_loc_taxa_cover!(relative_cover::AbstractArray{T,4}, out_relative_loc_taxa_cover::AbstractArray{T,3})::Nothing where {T<:Real}

Calculate the relative taxa cover for each location.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

out_relative_loc_taxa_cover : Output array buffer with dimensions [timesteps ⋅ groups ⋅ locations].

ADRIAIndicators.relative_loc_taxa_cover — Method

relative_loc_taxa_cover(relative_cover::AbstractArray{T,4})::AbstractArray{T,3} where {T<:Real}

Calculate the relative taxa cover for each location.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

Returns

A 3D array of relative taxa cover with dimensions [timesteps ⋅ groups ⋅ locations].

ADRIAIndicators.relative_taxa_cover! — Method

relative_taxa_cover!(relative_cover::AbstractArray{T,4}, location_area::AbstractVector{T}, out_relative_taxa_cover::AbstractArray{T,2})::Nothing where {T<:Real}

Calculate the relative taxa cover, summed up across all locations.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

location_area : The coral habitable area for each location.

out_relative_taxa_cover : Output array buffer with dimensions [timesteps ⋅ groups].

ADRIAIndicators.relative_taxa_cover — Method

relative_taxa_cover(relative_cover::AbstractArray{T,4}, location_area::AbstractVector{T})::AbstractArray{T,2} where {T<:Real}

Calculate the relative taxa cover, summed up across all locations.

Arguments

relative_cover : Relative cover with dimensions [timesteps ⋅ groups ⋅ sizes ⋅ locations].

location_area : The coral habitable area for each location.

Returns

A 2D array of relative taxa cover with dimensions [timesteps ⋅ groups].