Reef Indices

This section provides high-level reef indices. These indices combine multiple metrics to provide a summary of reef state for different purposes.

ADRIAIndicators.reef_biodiversity_condition_index! — Method

reef_biodiversity_condition_index!(rc::AbstractArray{T,2}, cd::AbstractArray{T,2}, sv::AbstractArray{T,2}, out_rbci::AbstractArray{T,2})::Nothing where {T<:AbstractFloat}

Calculate the Reef Biodiversity Condition Index (RBCI). This is implementation is limited to using outputs that can be provided by coral ecology models.

Arguments

rc : Relative coral cover.

cd : Coral diversity.

sv : Relative shelter volume.

out_rbci : Output array buffer for the RCI.

ADRIAIndicators.reef_biodiversity_condition_index — Method

reef_biodiversity_condition_index(relative_cover::AbstractArray{T,2}, coral_diversity::AbstractArray{T,2}, shelter_volume::AbstractArray{T,2})::AbstractArray{T,2} where {T<:AbstractFloat}

Calculate the Reef Biodiversity Condition Index (RBCI). The RBCI is simply the average of relative cover (RC), coral diversity (CD), and shelter volume (SV). Given as

\[\begin{align*} \text{RBCI} = \frac{\text{RC} + \text{CD} + \text{SV}}{3}. \end{align*}\]

This is implementation is limited to using outputs that can be provided by coral ecology models.

Arguments

relative_cover : Relative coral cover.

coral_diversity : Coral diversity.

shelter_volume : Relative shelter volume.

Returns

A 2D array of the Reef Condition Index.

ADRIAIndicators.reef_condition_index! — Method

reef_condition_index!(ltmp_cover::AbstractArray{T,2}, relative_shelter_volume::AbstractArray{T,2}, juvenile_indicator::AbstractArray{T,2}, rubble::AbstractArray{T,2}, out_rci::AbstractArray{T,2})::Nothing where {T<:AbstractFloat}

Calculate the Reef Condition Index (RCI).

This method uses four inputs: LTMP cover, relative shelter volume, juvenile indicator, and rubble. This is implementation is limited to using outputs that can be provided by ecological models that only model corals.

Arguments

ltmp_cover : LTMP Coral Cover with dimensions [timesteps ⋅ locations].

relative_shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

juvenile_indicator: Juvenile Indicator with dimensions [timesteps ⋅ locations].

rubble : Relative rubble cover with dimensions [timesteps ⋅ locations].

out_rci : Output RCI buffer with dimensions [timesteps ⋅ locations].

Reference

Ryan F. Heneghan, Gabriela Scheufele, Yves-Marie Bozec et al. A framework to inform economic valuation of non-use benefits from coral-reef intervention efforts, 02 October 2025, PREPRINT (Version 1) available at Research Square [https://doi.org/10.21203/rs.3.rs-7644150/v1]

ADRIAIndicators.reef_condition_index! — Method

reef_condition_index!(ltmp_cover::AbstractArray{T,2}, relative_shelter_volume::AbstractArray{T,2}, juvenile_indicator::AbstractArray{T,2}, out_rci::AbstractArray{T,2})::Nothing where {T<:AbstractFloat}

Calculate the Reef Condition Index (RCI).

This method uses three inputs: LTMP coral cover, relative shelter volume, and juvenile indicator. This is implementation is limited to using outputs that can be provided by ecological models that only model corals.

Arguments

ltmp_cover : LTMP Coral Cover with dimensions [timesteps ⋅ locations].

relative_shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

juvenile_indicator: Juvenile Indicator with dimensions [timesteps ⋅ locations].

out_rci : Output RCI buffer with dimensions [timesteps ⋅ locations].

Reference

Ryan F. Heneghan, Gabriela Scheufele, Yves-Marie Bozec et al. A framework to inform economic valuation of non-use benefits from coral-reef intervention efforts, 02 October 2025, PREPRINT (Version 1) available at Research Square [https://doi.org/10.21203/rs.3.rs-7644150/v1]

ADRIAIndicators.reef_condition_index — Method

reef_condition_index(relative_cover::AbstractArray{T,2}, shelter_volume::AbstractArray{T,2}, relative_juveniles::AbstractArray{T,2}, rubble::AbstractArray{T,2})::AbstractArray{T,2} where {T<:AbstractFloat}

Calculate the Reef Condition Index (RCI).

This method uses four inputs: LTMP cover, relative shelter volume, juvenile indicator, and rubble.

The RCI is a categorical index assessing the overall health and condition of a reef location based on five key ecological metrics. The index assigns a discrete score (0.1, 0.3, 0.5, 0.7, or 0.9) representing categories from "Very Poor" to "Very Good".

For each input there are five levels of condition ranging from very poor to very good. COTS and Rubble Cover is inverted, where high values indicate worse condition. Then the location is assigned a condition of very poor to very good if that location meets at least 2 of the metrics condition criteria. The condition level is then assigned a numerical value based on its categorisation.

Arguments

ltmp_cover : Relative Coral Cover with dimensions [timesteps ⋅ locations].

relative_shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

juvenile_indicator: Relative juvenile cover with dimensions [timesteps ⋅ locations].

rubble : Relative rubble cover with dimensions [timesteps ⋅ locations].

Returns

Output RCI buffer with dimensions [timesteps ⋅ locations].

Reference

Ryan F. Heneghan, Gabriela Scheufele, Yves-Marie Bozec et al. A framework to inform economic valuation of non-use benefits from coral-reef intervention efforts, 02 October 2025, PREPRINT (Version 1) available at Research Square [https://doi.org/10.21203/rs.3.rs-7644150/v1]

ADRIAIndicators.reef_condition_index — Method

reef_condition_index(ltmp_cover::AbstractArray{T,2}, relative_shelter_volume::AbstractArray{T,2}, juvenile_indicator::AbstractArray{T,2})::AbstractArray{T,2} where {T<:AbstractFloat}

Calculate the Reef Condition Index (RCI).

This method uses three inputs: LTMP cover, relative shelter volume, and juvenile indicator.

The RCI is a categorical index assessing the overall health and condition of a reef location based on three key ecological metrics. The index assigns a discrete score (0.1, 0.3, 0.5, 0.7, or 0.9) representing categories from "Very Poor" to "Very Good".

For each input there are five levels of condition ranging from very poor to very good. Then the location is assigned a condition of very poor to very good if that location meets at least two of the metrics condition criteria. The condition level is then assigned a numerical value based on its categorisation.

Arguments

ltmp_cover : Relative Coral Cover with dimensions [timesteps ⋅ locations].

relative_shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

juvenile_indicator: Relative juvenile cover with dimensions [timesteps ⋅ locations].

Returns

Output RCI buffer with dimensions [timesteps ⋅ locations].

Reference

Ryan F. Heneghan, Gabriela Scheufele, Yves-Marie Bozec et al. A framework to inform economic valuation of non-use benefits from coral-reef intervention efforts, 02 October 2025, PREPRINT (Version 1) available at Research Square [https://doi.org/10.21203/rs.3.rs-7644150/v1]

ADRIAIndicators.reef_fish_index! — Method

reef_fish_index!(rc::AbstractArray, out_rfi::AbstractArray)::Nothing

Calculate the Reef Fish Index (RFI) for a single scenario.

Arguments

rc : Relative coral cover with dimensions [timesteps ⋅ locations].

out_rfi : Output array buffer for the RFI with dimensions [timesteps ⋅ locations].

ADRIAIndicators.reef_fish_index — Method

reef_fish_index(relative_cover::AbstractArray{T,2},)::AbstractArray{T,2} where {T<:Real}

Calculate the Reef Fish Index (RFI) for a single scenario. The RFI is composed of two linear regressions mapping relative coral cover to structural complexity and finally, structural complexity to fish biomass. The index is based off figure 4a and 6b in Graham et al., 2013 [1]. RFI (kg/km²) as a functional of relative cover ($x$) is given as

\[\begin{align*} \text{SC}(x) &= 1.232 .+ 0.007476 ⋅ x ⋅ 100\\ \text{RFI} &= 0.01 ⋅ (-1623.6 + 1883.3 ⋅ SC), \end{align*}\]

where SC is the structural complexity of the location.

Arguments

relative_cover : Relative coral cover with dimensions [timesteps ⋅ locations].

Returns

A 2D array of the Reef Fish Index with dimensions [timesteps ⋅ locations].

References

Graham, N.A.J., Nash, K.L. The importance of structural complexity in coral reef ecosystems. Coral Reefs 32, 315–326 (2013). https://doi.org/10.1007/s00338-012-0984-y

ADRIAIndicators.reef_tourism_index! — Method

reef_tourism_index!(relative_cover::AbstractArray{T,2}, shelter_volume::AbstractArray{T,2}, relative_juveniles::AbstractArray{T,2}, cots::AbstractArray{T,2}, rubble::AbstractArray{T,2}, out_rti::AbstractArray{T,2})::Nothing where {T<:AbstractFloat}

Calculate the Reef Tourism Index (RTI) for a single scenario.

This method uses five inputs: relative cover, shelter volume, relative juveniles, COTS abundance, and rubble.

Arguments

relative_cover : Relative coral cover with dimensions [timesteps ⋅ locations].

shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

relative_juveniles : Relative juvenile cover with dimensions [timesteps ⋅ locations].

cots : COTS abundance as a count with dimensions [timesteps ⋅ locations].

rubble : Rubble as a proportion of location area with dimensions [timesteps ⋅ locations].

out_rti : Output array buffer for the RTI with dimensions [timesteps ⋅ locations].

ADRIAIndicators.reef_tourism_index! — Method

reef_tourism_index!(relative_cover::AbstractArray{T,2}, coral_evenness::AbstractArray{T,2}, relative_shelter_volume::AbstractArray{T,2}, relative_juveniles::AbstractArray{T,2}, out_rti::AbstractArray{T,2})::Nothing where {T<:AbstractFloat}

Calculate the Reef Tourism Index (RTI) for a single scenario.

This method uses four inputs: relative cover, coral evenness, relative shelter volume, and relative juveniles.

Arguments

relative_cover : Relative coral cover with dimensions [timesteps ⋅ locations].

coral_evenness : Coral evenness with dimensions [timesteps ⋅ locations].

relative_shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

relative_juveniles : Relative juvenile cover with dimensions [timesteps ⋅ locations].

out_rti : Output array buffer for the RTI with dimensions [timesteps ⋅ locations].

ADRIAIndicators.reef_tourism_index — Method

reef_tourism_index(relative_cover::AbstractArray{T,2}, shelter_volume::AbstractArray{T,2}, relative_juveniles::AbstractArray{T,2}, cots::AbstractArray{T,2}, rubble::AbstractArray{T,2})::AbstractArray{T,2} where {T<:AbstractFloat}

Calculate the Reef Tourism Index (RTI) for a single scenario. The RTI is the Reef Condition Index made continuous by fitting a linear regression model using relative cover, shelter volume, relative juveniles, cots abundance and rubble to underpin it.

This method uses five inputs: relative cover, shelter volume, relative juveniles, COTS abundance, and rubble.

Arguments

relative_cover : Relative Coral Cover with dimensions [timesteps ⋅ locations].

shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

relative_juveniles : Relative juvenile cover with dimensions [timesteps ⋅ locations].

cots : COTS abundance with dimensions [timesteps ⋅ locations].

rubble : Rubble as proportion of location area with dimensions [timesteps ⋅ locations].

Returns

A 2D array of the Reef Tourism Index with dimensions [timesteps ⋅ locations].

ADRIAIndicators.reef_tourism_index — Method

reef_tourism_index(relative_cover::AbstractArray{T,2}, coral_evenness::AbstractArray{T,2}, relative_shelter_volume::AbstractArray{T,2}, relative_juveniles::AbstractArray{T,2})::AbstractArray{T,2} where {T<:AbstractFloat}

Calculate the Reef Tourism Index (RTI) for a single scenario.

This method uses four inputs: relative cover, coral evenness, relative shelter volume, and relative juveniles.

The RTI is a continuous version of the Reef Condition Index, fitted with a linear regression model.

Arguments

relative_cover : Relative coral cover with dimensions [timesteps ⋅ locations].

coral_evenness : Coral evenness with dimensions [timesteps ⋅ locations].

relative_shelter_volume : Relative shelter volume with dimensions [timesteps ⋅ locations].

relative_juveniles : Relative juvenile cover with dimensions [timesteps ⋅ locations].

Returns

A 2D array of the Reef Tourism Index with dimensions [timesteps ⋅ locations].